requirements.txt
python==3.8.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.7.1
Jinja2==3.1.4
assemblyai==0.21.0
google-generativeai==0.5.4
